+++ BomberGirl/pick_up_Character_state.h
@@ -40,7 +40,7 @@ namespace Bombergirl{
         void update(const float&) override;
         void render() override;
         void lockSelect(int, sf::RectangleShape&, const sf::Vector2f& , const sf::Vector2f&);
-        void setText(sf::Text &text, const float&, const sf::String&);
+        void setText(sf::Text &text, const int&, const sf::String&);
 	};
 }
